+++ src/Model/ProviderRepository.php
@@ -18,9 +18,12 @@ final class ProviderRepository extends ContaoRepository
      * @param int[] $providerIds
      *
      * @return ProviderModel[]|Collection|null
+     * @psalm-return Collection|null
+     * @psalm-suppress MixedReturnStatement
+     * @psalm-suppress MixedInferredReturnType
      */
     public function findByIds(array $providerIds) : ?Collection
     {
-        return $this->findMultipleByIds($providerIds);
+        return $this->__call('findMultipleByIds', [$providerIds]);
     }
 }