+++ src/Provider/ConfigProvider.php
@@ -6,6 +6,9 @@ namespace Cowegis\ContaoGeocoder\Provider;
 
 use Traversable;
 
+/**
+ * @extends Traversable<array{type: string, title: ?string, id: string}>
+ */
 interface ConfigProvider extends Traversable
 {
 }