+++ src/Provider/ConfigProvider/ConfigProviderChain.php
@@ -8,21 +8,23 @@ use AppendIterator;
 use Cowegis\ContaoGeocoder\Provider\ConfigProvider;
 use IteratorAggregate;
 use IteratorIterator;
+use Traversable;
 
+/** @implements IteratorAggregate<array{type: string, title: ?string, id: string}> */
 final class ConfigProviderChain implements IteratorAggregate, ConfigProvider
 {
-    /** @var ConfigProvider[] */
+    /** @var list<ConfigProvider> */
     private $configProviders;
 
     /**
-     * @param ConfigProvider[] $configProviders
+     * @param list<ConfigProvider> $configProviders
      */
     public function __construct(iterable $configProviders)
     {
         $this->configProviders = $configProviders;
     }
 
-    /** @return mixed[][] */
+    /** @return Traversable<array{type: string, title: ?string, id: string}> */
     public function getIterator() : iterable
     {
         $iterator = new AppendIterator();