+++ src/Provider/ConfigProvider/DatabaseConfigProvider.php
@@ -5,14 +5,17 @@ declare(strict_types=1);
 namespace Cowegis\ContaoGeocoder\Provider\ConfigProvider;
 
 use ArrayIterator;
+use Contao\CoreBundle\Framework\ContaoFramework;
 use Cowegis\ContaoGeocoder\Model\ProviderRepository;
 use Cowegis\ContaoGeocoder\Provider\ConfigProvider;
 use Cowegis\ContaoGeocoder\Provider\ProviderFactory;
 use IteratorAggregate;
+use Traversable;
 use function count;
 use function sprintf;
 use function str_repeat;
 
+/** @implements IteratorAggregate<array{type: string, title: ?string, id: string}> */
 final class DatabaseConfigProvider implements IteratorAggregate, ConfigProvider
 {
     /** @var ProviderRepository */
@@ -21,27 +24,36 @@ final class DatabaseConfigProvider implements IteratorAggregate, ConfigProvider
     /** @var ProviderFactory */
     private $providerFactory;
 
-    public function __construct(ProviderRepository $repository, ProviderFactory $providerFactory)
-    {
+    /** @var ContaoFramework */
+    private $framework;
+
+    public function __construct(
+        ProviderRepository $repository,
+        ProviderFactory $providerFactory,
+        ContaoFramework $framework
+    ){
         $this->repository      = $repository;
         $this->providerFactory = $providerFactory;
+        $this->framework       = $framework;
     }
 
-    /** @return mixed[][] */
-    public function getIterator() : iterable
+    /** @return Traversable<array{type: string, title: ?string, id: string}> */
+    public function getIterator() : Traversable
     {
         $types = $this->providerFactory->typeNames();
         if (count($types) === 0) {
             return new ArrayIterator();
         }
 
+        /** @psalm-suppress InternalMethod */
+        $this->framework->initialize();
         $collection = $this->repository->findBy(
             [sprintf('.type IN (?%s)', str_repeat(',?', count($types) - 1))],
             $types,
-            ['.isDefault \'1\',\'\'']
+            ['order' => 'FIELD( .isDefault, \'1\',\'\')']
         );
 
-        if ($collection) {
+        if ($collection !== null) {
             return new ArrayIterator($collection->fetchAll());
         }