+++ src/Provider/Geocoder.php
@@ -11,8 +11,10 @@ use Geocoder\Provider\Provider as GeocodeProvider;
 use Geocoder\Query\GeocodeQuery;
 use Geocoder\Query\ReverseQuery;
 use IteratorAggregate;
+use Traversable;
 use function array_keys;
 
+/** @implements IteratorAggregate<Provider> */
 final class Geocoder implements GeocodeProvider, IteratorAggregate
 {
     /** @var Provider[] */
@@ -39,8 +41,8 @@ final class Geocoder implements GeocodeProvider, IteratorAggregate
         throw ProviderNotRegistered::create($providerId, array_keys($this->providers));
     }
 
-    /** @return Provider[] */
-    public function getIterator() : iterable
+    /** @retun Traversable<Provider> */
+    public function getIterator() : Traversable
     {
         return new ArrayIterator($this->providers);
     }