+++ src/Provider/GeocoderProviderDecoratorFactory.php
@@ -7,9 +7,15 @@ namespace Cowegis\ContaoGeocoder\Provider;
 use Assert\Assert;
 use Geocoder\Provider\Provider as GeocodeProvider;
 
+/**
+ * @psalm-type TProviderConfig = array{title: ?string, id: string }
+ */
 trait GeocoderProviderDecoratorFactory
 {
-    /** @param mixed[] $config */
+    /**
+     * @param mixed[] $config
+     * @psalm-param TProviderConfig $config
+     */
     protected function createDecorator(GeocodeProvider $provider, array $config) : Provider
     {
         Assert::that($config)