+++ src/Provider/ProviderType/ChainProviderFactory.php
@@ -5,11 +5,17 @@ declare(strict_types=1);
 namespace Cowegis\ContaoGeocoder\Provider\ProviderType;
 
 use Contao\StringUtil;
+use Cowegis\ContaoGeocoder\Model\ProviderModel;
 use Cowegis\ContaoGeocoder\Model\ProviderRepository;
 use Cowegis\ContaoGeocoder\Provider\Provider;
 use Cowegis\ContaoGeocoder\Provider\ProviderFactory;
 use Geocoder\Provider\Chain\Chain;
+use function array_map;
+use function assert;
 
+/**
+ * @psalm-type TChainConfig = array{type: string, title: ?string, id: string, chain_providers?: string }
+ */
 final class ChainProviderFactory extends BaseProviderTypeFactory
 {
     protected const FEATURES = [Provider::FEATURE_ADDRESS, Provider::FEATURE_REVERSE];
@@ -31,14 +37,23 @@ final class ChainProviderFactory extends BaseProviderTypeFactory
         return 'chain';
     }
 
-    /** {@inheritDoc} */
-    public function create(array $config) : Provider
+    /**
+     * {@inheritDoc}
+     * @psalm-param TChainConfig $config
+     * @psalm-suppress MoreSpecificImplementedParamType
+     */    public function create(array $config) : Provider
     {
         $chain       = new Chain();
-        $providerIds = StringUtil::deserialize($config['chain_providers'] ?? '', true);
-        $providers   = $this->repository->findByIds($providerIds);
+        $providerIds = array_map(
+            'intval',
+            (array) StringUtil::deserialize($config['chain_providers'] ?? '', true)
+        );
+
+        $providers = $this->repository->findByIds($providerIds);
 
         foreach ($providers ?: [] as $provider) {
+            assert($provider instanceof ProviderModel);
+            /** @psalm-suppress MixedArgumentTypeCoercion */
             $chain->add($this->factory->create($provider->type, $provider->row()));
         }