+++ CHANGELOG.md
@@ -1,7 +1,17 @@
 # Changelog
 
+## [0.2.0] - 2020-12-08
+
+### Added
+
+ - Provide search api endpoint and allow to protect it with an api key and referrer check
+ - Allow to limit nominatim geocoder to specific country codes
+
 ## dev-master
 
 ### Removed
 
  - Remove scope definition
+
+
+[0.2.0]: https://github.com/cowegis/contao-geocoder/compare/0.1.0...0.2.0